Changed cookie expiry date to a year instead of 6 minutes (#2)
## Type of change
##### ```Select the type of change ```
+ #### [ ] Frontend
+ #### [x] Backend
+ #### [ ] Database
+ #### [ ] Utilities

## Feature description
##### ```Clearly and conscisely describe the feature.```
Changed cookies & sessions exipiry date to be one year instead of 6 minutes.

## Solution description
##### ```Describe the code changes in detail.```
First, i modified `cookies.js`. That is the responsible file to be creating cookies on the frontend.
setTime() takes in time in milliseconds, so the appropiate forumala to calculate 1 year in milliseconds would
be 1000 (milliseconds in second) * 3600 (seconds in an hour) * 24 (hours in a day) * 365 (days in a year).

Then i changed `authentication.go`. That is reponsible for creating the session in `database.db`.
SQL has a query parameter `datetime`. With that combined with its own parameters you are able to
add which ever time to datetime('now'). In our case datetime('now','+1 years').

## Output screenshots
###### ```Post output screenshots, if an UI is affected or added to this feature.```
None

## Tested browsers
##### ```Please select the brwosers you tested the feature it is an UI change.```
+ #### [x] Chrome
+ #### [x] Firefox
+ #### [ ] Edge
+ #### [ ] Safari
+ #### [ ] Not needed

Co-authored-by: nimi25820 <robert.laur.private@gmail.com>
Reviewed-on: https://01.kood.tech/git/nimi25820/social-network/pulls/2
Reviewed-by: Gunta Kļava <gukka@noreply.01.kood.tech>
Co-authored-by: nimi25820 <nimi25820@noreply.01.kood.tech>
Co-committed-by: nimi25820 <nimi25820@noreply.01.kood.tech>