profile-overhaul (#5)
## Type of change
##### ```Select the type of change ```
+ #### [x] Frontend
+ #### [x] Backend
+ #### [x] Database
+ #### [ ] Utilities

## Feature description
##### ```Clearly and conscisely describe the feature.```

Profiles are publicly available to be viewed by others and are not only viewable by the owner.

Added public/ private profile. User is able to select between public/private mode. Public profiles are available for everyone whereas private profiles requires user1 to be following user2 in order for user1 to see user2 profile.

Added follow user functionality. If profile is public, you can instantly start following a user, but if it is private, you first have to request user2 to start following. User2 can then accept or deny.

Added image support to profile page. Now you can see your profile image if you chose to upload one upon registration.

Added Followers section to profile page. There it is possible to see Followers, Following and if the owner is fetching the page, then Follower Requests. There you can perform various
functionalities like removing followers, accepting followers or unfollowing users.

## Solution description
##### ```Describe the code changes in detail.```

./profile/{user}

Upon loading a profile. A request to fetch profile data will be made. (func Profile) The first step is to check if the profile actually exists, so i query the Users table to check if the user exists. If it doesnt exist, we send back profile doesnt exist.

Then i check if the person wanting to see a profile is actually the owner of profile. That i achvie by comparing name gotten with cookie token and the profile wanted to see. If they are equal, that means that it is an owner. We need this to send back a status = "owner". I use this status to display certain frontend elements.

Then if the person wanting to see a profile is not the owner of the profile, i check if the profile is public or private. That i get from Users table profile_status. If the profile is public we can send back the data instantly but i stiil check if the person is following or not, to display the correct button either Follow/Unfollow on the frtonend.

If the profile is private, then we also have to check, if the person wanting to see a profile, is a accepted follower of wanted users profile. I made a function CheckFollower. Function returns a boolean. It just checks if Followers table has a row where status "following", follower_id = user wanting to see a profile, recipient_id = user whose profile is trying to be fetched.

./profile/{user}/followers

Everything is basically the same compared to fetching profile page. Just the data that is being fetched changes.

New functionalities will be Unfollow button, Remove button and accept/deny button if it is the owner viewing its own page.

With both pages, if the profile being viewed is private and the person wanting to view the profile is not a follower. A prompt giving the person to request to follow a person is given. The person wont be able to see the profile until the owner of the profile has accepted the follow request.

## Output screenshots
###### ```Post output screenshots, if an UI is affected or added to this feature.```
https://gyazo.com/ee2adb52b437aea11e0af7d3f530cd75;
https://gyazo.com/7851c46a641af8c92520ef168f65d0ce;
https://gyazo.com/8682263cf48bca9cb112a3b04c9012be;

## Tested browsers
##### ```Please select the brwosers you tested the feature it is an UI change.```
+ #### [x] Chrome
+ #### [x] Firefox
+ #### [x] Edge
+ #### [ ] Safari
+ #### [ ] Not needed

Co-authored-by: gukka <guka245@gmail.com>
Co-authored-by: nimi25820 <robert.laur.private@gmail.com>
Reviewed-on: https://01.kood.tech/git/nimi25820/social-network/pulls/5
Reviewed-by: Gunta Kļava <gukka@noreply.01.kood.tech>
Co-authored-by: nimi25820 <nimi25820@noreply.01.kood.tech>
Co-committed-by: nimi25820 <nimi25820@noreply.01.kood.tech>